[DX] Extract JobInterface and RuntimeErrorException
