+++ spec/Pim/Bundle/ImportExportBundle/Validator/Constraints/JobInstanceValidatorSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\ImportExportBundle\Validator\Constraints;
 
 use Akeneo\Bundle\BatchBundle\Connector\ConnectorRegistry;
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
-use Akeneo\Bundle\BatchBundle\Job\JobInterface;
+use Akeneo\Component\Batch\Job\JobInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\ImportExportBundle\Validator\Constraints\JobInstance as JobInstanceConstraint;
 use Prophecy\Argument;
