+++ src/Pim/Bundle/BaseConnectorBundle/Writer/File/CsvProductWriter.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\BaseConnectorBundle\Writer\File;
 
-use Akeneo\Bundle\BatchBundle\Job\RuntimeErrorException;
+use Akeneo\Component\Batch\Job\RuntimeErrorException;
 use Akeneo\Component\FileStorage\Exception\FileTransferException;
 use Pim\Bundle\BaseConnectorBundle\Writer\File\ContextableCsvWriter;
 use Pim\Component\Connector\Writer\File\FileExporterInterface;
