+++ src/Pim/Bundle/BaseConnectorBundle/Writer/File/FileWriter.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\BaseConnectorBundle\Writer\File;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Component\Batch\Item\AbstractConfigurableStepElement;
 use Akeneo\Component\Batch\Item\ItemWriterInterface;
-use Akeneo\Bundle\BatchBundle\Job\RuntimeErrorException;
+use Akeneo\Component\Batch\Job\RuntimeErrorException;
 use Akeneo\Component\Batch\Step\StepExecutionAwareInterface;
 use Pim\Bundle\ImportExportBundle\Validator\Constraints\WritableDirectory;
 use Symfony\Component\Validator\Constraints as Assert;
